Update version on upgrade script 06
This script was created while USM major release deployment
tests were being run between stx-9 -> stx-10 releases. Now that
the master branch was upversioned to stx-10, this script needs
to be updated, otherwise it will not run during "migrate" step.

If this script does not run, the deploy host phase fails to set
the target load for the hosts, and the hieradata is not generated
correctly to the host fails configuration after the unlock.

Test Plan
PASS: verify that script runs on stx-10 -> fake stx-11 deployment
PASS: deploy all hosts with stx-11 in a standard deployment

Story: 2010676
Task: 50400

Change-Id: I182ae221401c09e8eee7054ad07e0d3a7c6f942f
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>