Accept management and multicat secondary pool config at runtime
This changes adds the possibility to add (and remove) a secondary pool
for the management and multicast networks. A separated task will be
added later to trigger a "configuration out-of-date" alarm to inform
the operator that a node lock/unlock is required, in a similar way
that is done to the OAM network.

The management network applies the floating and static node addresses
after lock/unlock. For multicast it just stores in sysinv's database
without a system usage as this type of address is used mainly for
maintenance and patching, and they use the primary pool addresses.

Test Plan:

The test cases added a secondary pool after installation and executed
a lock/unlock on the nodes to check that the mgmt configuration was
applied.

[PASS] install an IPv6 AIO-DX
[PASS] install an IPv6 AIO-SX
[PASS] install an IPv4 AIO-DX
[PASS] install an IPv4 AIO-SX
[PASS] unit tests

Story: 2011027
Task: 50423

Change-Id: I9ff16f5e7c6da078bba690e47a15756436f0d108
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>