Host-unlock: delete and recreate k8s cluster-info configmap
Procedure to add a new node to a kubernetes cluster involves
creating a new kubeadm token and using it when a node is joined.
After kubeadm token is created, kubernetes updates the cluster-info
configmap with the token. "kubeadm join" command consumes this
JWS signed token from the cluster-info configmap to join a node to
the cluster.

There is a rare bug upstream without solution that causes
cluster-info configmap to be not updated after token creation.

This implements an upstream workaround (https://discuss.kubernetes.io/t/cluster-info-configmap-does-not-yet-contain-a-jws-signature/14985)
to retrieve the cluster-info configmap, delete the exisitng version
and create a new version from the retrieved one without any
modifications.

To mitigate a rare condition of the kube-apiserver being down after
retrieving and deleting the cluster-info configmap and before
recreating the same, a retry decorator is added over the method that
regenerates the configmap. If it still fails, cluster-info configmap
can be manually regenerated from the file system location
/var/run/kubernetes/cluster-info.yaml.

Although it was preferable to use kubernetes python library methods
to retrieve, delete and recreate the configmap, kubectl calls were
preferred for easier dumping of cluster-info configmap on the
file-system for the cluster-info recovery if anything goes wrong.

Test Plan:
PASS: AIO-SX installation successful.
PASS: Storage lab (2+2+1) installation successful.
PASS: AIO-DX upgrade and downgrade successful.

Closes-Bug: 2070785

Change-Id: I787a521a31deffbda14813c399762067baf8ad31
Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>