Fix support for apps with multiple namespaces
Helm allows charts to span multiple namespaces. However, the
behavior implemented in the Application Framework was using only
namespaces available on kustomization.yaml files to look for enabled
charts. This strategy breaks the multiple namespace support because
kustomize requires each HelmRelease resource to have a different
namespace to support this scenario. Using a namespace on the a
kustomization.yaml file means that a single namespace would be used
across all underlying resources, including Helm releases.

As a result, the "is_chart_enabled" function, would always return no
charts for the multiple namespaces scenario, since the release
namespace and the namespace used by the framework would not match,
causing Helm overrides not to work correctly.

The solution consists of using the output of the "kubectl kustomize
<fluxcd_directory>" command and checking the namespace present
within each HelmRelease. That way we ensure that the HelmRelease
manifest always has a namespace, whether it is inherited by a
kustomization.yaml file or explicitly defined on the release manifest.

Test Plan:
PASS: Upload, apply, remove and delete platform-integ-app
PASS: Upload, apply, remove and delete metrics-server
PASS: Upload, apply, remove and delete Kubevirt
PASS: Update user overrides to metrics-server and platform-integ-app
      via "system helm-override-update" and when reapplying the app,
      the changes were made successfully
PASS: Update the user overrides of the kubevirt app for the kubevirt
      namespace: "system helm-override-update kubevirt-app
      kubevirt-app kubevirt --set replicas=3". After running the
      command and reapplying the app, the changes were made
      successfully.
PASS: Update the user overrides of the kubevirt app for the cdi
      namespace: "system helm-override-update kubevirt-app
      kubevirt-app cdi --set replicas=3". After running the command
      and reapplying the app, the changes were made successfully.

Closes-bug: 2071380

Change-Id: I8a7535a4303f3a64d99cc8cc3db36bd2d6d6ba5d
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>