Fix k8s upgrade abort during pre app update step
This commit fixes a race condition when a 'kube-upgrade-abort' command
is issued while the Kubernetes upgrade is in 'pre-updating-apps' state.
The upgrade status is correctly changed to 'upgrade-aborted' at first,
but later on it is mistakenly changed to 'pre-updated-apps' or
'pre-updating-apps-failed', depending on the application update result.

A check was introduced to verify whether the Kubernetes upgrade process
is being aborted. If so, the upgrade remains in the resulting abort
state, thus respecting the outcome of the 'kube-upgrade-abort' command.

Test Plan:
PASS: Run a Kubernetes upgrade.
      Abort upgrade during the pre-updating-app state.
      Confirm that the upgrade was aborted and that the status does not
      change when we are done updating the apps.
PASS: Run a Kubernetes upgrade.
      Abort upgrade during the post-updating-app state.
      Confirm that the abort request was rejected.
PASS: Run a Kubernetes upgrade with apps to be updated during pre and
      post update phases.
      Confirm that the upgrade was successful.
PASS: Run a Kubernetes upgrade without any application updates
      Confirm that the upgrade was successful.

Closes-bug: 2071743

Change-Id: I6b34e8e8cfd28395d8b7e083eabd58e83c13eaa9
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>