Trigger "out-of-date" alarm for mgmt secondary pool runtime config
The management network allows adding or removing a secondary pool
configuration at runtime. However, this change requires an operator to
perform a lock/unlock operation on the nodes for the new configuration
to take effect.

To initiate this notification, the basic network and firewall
configuration are applied. The generated configuration is then flagged
for a system reboot. This ensures that any services relying on this
new network address family remain available after the reboot.

This change also triggers runtime storage network configuration. This
network is used for external NetApp deployment as the storage backend.
It was deemed not necessary to trigger an alarm for this one as all
configuration needed (adresses, interfaces and firewall) can be at
runtime. The trident service will be available via pods, not requiring
a node restart.

Test Plan
- management
[PASS] install STANDARD, then add a secondary management network verify
       that a "config out-of-date" is raised on all affected nodes.
[PASS] execute lock/swact/unlock for controllers ans lock/unlock for
       workers and check that the alarm is removed
[PASS] verify linux config (addresses and firewall) for the new
       network on each node.
-storage
[PASS] with a STANDARD system add the network and configure address
       pools and interfaces to this network and check if they are
       applied to the nodes as the configuration is executed.

Story: 2011027
Task: 50500

Change-Id: I564cc72d6b5d93fad1956e79903265aa9f62efcb
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>