Fix upgrade script (#41) for SX USM upgrade
Some steps done by the ansible upgrade playbook in legacy upgrade
to update the platform certificates and system-local-ca data for the
new configuration aren't being executed now that USM is implemented.

These steps were already implemented in an upgrade script (#41), since
they were required for legacy DX upgrade. The script was adapted to handle SX as well.

Include extra flags in the user config folder, signaling that the
data for system-local-ca was provided by user (in bootstrap or via the
update platform certificates playbook) or that the certificates were
updated during an upgrade.

With this, we can verify in the health check for a future upgrade if
the user has provided the recommended data for system-local-ca.

Test plan:
PASS: Deploy AIO-SX w/ lab config for the platform certificates.
      Verified that the upgrade health check fails.
PASS: Deploy AIO-SX w/ production config for the platform certificates.
      Verified that the upgrade health check passes.
PASS: Upgraded AIO-SX from stx 9 to stx 10. Verified that the upgrade
      health check passes (upgraded systems are expected to be
      migrated).

Depends-on: https://review.opendev.org/c/starlingx/ansible-playbooks/+/923902

Story: 2009811
Task: 50550

Change-Id: I9c0fb279460f0a6fde932e7ba378dc05738e50f2
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>