Update software_version in db during the upgrade
This commit is to update the software_version field in i_system
table during platform upgrade. This script is executed during
the activate or activate-rollback stage.

The software_version field can only be updated directly in database
operation. This is the same as what the legacy upgrade uses.

Test Plan:

PASS: build and deploy the iso on SX
PASS: manually run the script

Task: 50555
Story: 2010676
Change-Id: I3e901b780766d7bad3b6b05c00adefb99defbbd5
Signed-off-by: junfeng-li <junfeng.li@windriver.com>