Fix error condition for cert private key
A small logic error when checking the private key type for ECC or
RSA is causing the platform certificates to not be installed
properly by cert-mon.

Fix the logic condition to throw only if the private key of the
installed certificate is not ECC AND not RSA.

Test plan:
PASS: Bootstrap AIO-DX.
PASS: After unlocking c0, verify in the cert-mon logs that the platform
      certificates were installed correctly by sysinv API.
PASS: Verify that the ssl cert (server-cert.pem) is in the platform
      config folder.
PASS: Install and unlock the c1.

Story: 2009811
Task: 50584

Change-Id: I6dd7c3c8c3a9e60681d1cf7c279cf61fd8a4bb13
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>