Fix addrpool-modify for mgmt pools in AIO-SX when no-proxy list exists
addrpool-modify was causing an exception when a management pool was
modified in AIO-SX hosts which had a docker no-proxy list. This change
fixes the issue and adds an automated test to cover the scenario. Also,
functions related to the no-proxy list were moved to
common.address_pool since other controllers also depend on them.

Test plan
=========

Setup:
 - Management network assigned to mgmt0 ethernet interface
 - Pool mgmt-ipv4 with floating=192.168.204.1, c0=192.168.204.2
 - Pool mgmt-ipv6 with floating=2001::1, c0=2001::2
 - No-proxy list: 192.168.204.1,192.168.204.2,[2001::1],[2001::2]

[PASS] Modify mgmt-ipv4, floating=192.168.204.11, c0=192.168.204.12,
       check that no-proxy list was updated accordingly
[PASS] Modify mgmt-ipv6, floating=2001::11, c0=2001::12, check that
       no-proxy list was updated accordingly

Story: 2011027
Task: 50586
Change-Id: I0dae6d120b776d2c35e8757460d5d50e989b415d
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>