Improve SM audit of sysinv-inv performance
This change introduces support for the OPTIONS HTTP method
in the root controller of the sysinv-api.

Additionally, the OCF monitor function is updated to send
an HTTP OPTIONS request using `curl` to check the health
of the sysinv-api service. This ensures that the service
is properly monitored and health checks are performed
reliably.

To facilitate this, the sysinv-api now supports an
unauthenticated endpoint using the OPTIONS method, allowing
for health checks without requiring keystone authentication.

Test Plan:
[PASS] Verified that the OPTIONS method correctly lists the
allowed methods
[PASS] Confirmed that the OPTIONS method responds with HTTP
204 No Content, indicating successful handling of the request.
[PASS] Verify simulate fail path by making sysinv-api non-responsive
to OPTIONS request
[PASS] Verify system host-lock/unlock and system host-swact
[PASS] no failed services (fm alarm-list; sudo sm-dump)
[PASS] Verify sm-restart
[PASS] Verify monitor service works fine in IPV6 lab

Story: 2011106
Task: 50614

Change-Id: Ibe97359de40a60db8d2d84fac8200443c77b7c85
Signed-off-by: rummadis <ramu.ummadishetty@windriver.com>