Reinstate SSSD user/group objectSid based ID mapping
A change in the approach of how WAD users and groups are mapped
to Linux groups prompted this change in the SSSD configuration.
In the current implementation the mapping of WAD user and group
IDs is based on Posix schema IDs such as uidNumber and gidNumber.
This approach comes with extra work on the WAD server side to
create sudo rules for the WAD users/groups that are going to be
awarded sudo privileges. In addition, WAD users and groups
need to be configured with uidNumber and gidNumber in order to
have their group membership mapped to a Linux group.
A simplified approach is to use Linux pam_group configuration to
map/bind LDAP user and groups to Linux groups in order to award
stx platform Roles and Permissions to LDAP users/groups.
The new approach that uses the Linux pam_group configuration,
requires this commit to reinstate the former default SSSD
configuration, where the WAD users/groups are mapped to Linux
users/groups on stx platform, using Windows Security Identifiers
(objectSid).

Test Plan:
PASS: Successful install in AIO-SX system configuration.
PASS: The WAD users and groups have been mapped successfully
using ActiveDirectory objectSID mapping. No ID setting was required
on the WAD server
PASS: Add pam_group.so configuration in /etc/pam.d/common-auth.
Add a line in /etc/security/group.conf to map users of a WAD group
to a list of stx platform groups, e.g. "sys_protected", "sudo" and
"root". Login with a ldap user in the mapped WAD group and verify it
has membership in all the stx platform groups in the list.
PASS: Verify sudo capabilities of ldap users in the WAD group mapped
to the stx platform "sudo" group.
PASS: Verify WAD users mapped to the stx platfrom "sys_protected"
group can perform privileged operations like
"source /etc/platform/openrc".
PASS: Verify any WAD user has access to kubectl.
PASS: Add a line in /etc/security/group.conf to map users of the
local OpenLDAP group to a list of stx platform groups, e.g.
"sys_protected", "sudo" and "root". Login with a ldap user of the
mapped local OpenLDAP group and verify it has membership in all the
stx platform groups in the list.
Verify sudo, root and sys_protected privileges of an ldap user
added to those stx platform groups.

Story: 2011180
Task: 50608

Signed-off-by: Carmen Rata <carmen.rata@windriver.com>
Change-Id: Iff03896de2e9e6aae80b4d991839b82458b71b41