Fix kube-root-ca-update after rehoming DX sc
After performing a DX subcloud rehoming, updating the K8S RCA was
failing due to the pods scheduled in c1 not being able to pull
images from the local registry.

Investigation showed that the cause was the secrets containing the
auth data for user 'sysinv' not being updated after the keystone
password is changed during rehoming.

There is an audit task that runs once a day and updates the auth
data inside the secrets. This task is also triggered by a callback
when keystone password changes for the user 'admin'. This review adds
the callback for the user 'sysinv' as well.

Test plan:
PASS: Deploy DC + DX subcloud
      Rehome DX subcloud and perform kube-root-ca-update

Closes-bug: 2074386

Change-Id: Ia42e2a33342374d5e2c98b9b95b0713595783aa1
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>