Trigger runtime manifest during enrollment
During subcloud enrollment, the openstack runtime manifest are
expected to update the service configurations on a non-DC system(not
yet to be updated as a subcloud). This commit allows the puppet
runtime manifest to be applied after receiving the notification from
keystone listener after the service user's passwords are updated in
keystone.

Test plan:
Passed - Build an ISO with this change and bootstrap a AIO-SX and DC
system controllers.
Passed - Verified the keystone users' passwords are updated by puppet
runtime manifest after updating in keystone.

Story: 2011100
Task: 50666

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I01373fee403165c1396387e60732f49fe9a76fa2