Fix multiple datanetworks to interface restriction
API does not allow a second datanetwork assignment to an interface that
already has a datanetwork assigned. This limitation should just be
applied to SR-IOV interfaces since SR-IOV plugin v3.5.1 does not allow
that anymore.

Testplan
========

[PASS] Assign a SRIOV interface and datanetwork that already exists.
Configuration should not be allowed.
[PASS] Assign multiple datanetworks to same SRIOV interface.
Configuration should not be allowed.
[PASS] Assign multiple datanetworks to one data interface. Configuration
must be allowed.
[PASS] Assign multiple data interfaces to one datanetwork. Configuration
should not be allowed.
[PASS] Assign a data interface to datanetwork that already exists.
Configuration should not be allowed.
[PASS] Assign multiple datanetworks to one pci-passthrough interface.
configuration must be allowed.
[PASS] Assign a pci-passthrough interface to datanetwork that already
exists. Configuration should not be allowed.

Closes-Bug: 2075961

Change-Id: If7aa6f2c3dc148761b09d0b4d0ee4ea1efccf8bc
Signed-off-by: Ferdinando Terada <ferdinando.godoyterada@windriver.com>