Remove dcagent user creation in SystemController
It's not necessary to have the dcagent keystone user present on
system controller. Its presence can cause problems during upgrade
as it would be necessary for the dcagent user in the subcloud to
have the same password as system controller upon creation (before
dcdbsync synching upon managing the subcloud).

Test plan:
  - PASS: Build a developer ISO. Deploy a DC system with 2 system
          controllers and 1 subcloud. Verify the dcagent user is
          created only on the subcloud.
  - PASS: Manage the subcloud and verify the dcagent service is
          working as expected.
  - PASS: Verify dcagent user, service and endpoints are created
          during bootstrap.

Story: 2011106
Task: 50748

Change-Id: I4b52c6f322104c067d5cc568751eb94fc3fb3d7f
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>