Improve creation/deletion of controller-fs filesystem
This commit improves the creation/deletion of controller-fs
logic. Before, it was possible to create/delete the controllerfs
in some specific situations in which the standby controller is
not locked/online but passes the verification, which may impact
the manifests execution and the transition of states. This
improvement reinforces the logic of only being possible to
create/delete the controllerfs with the standby controller
locked. The tests were also updated to better reflect the
changes made.

Test Plan:
PASS: AIO-DX -> Both controllers unlocked-available +
      controllerfs add/delete + verify that the execution
      was interrupted and the error message was correctly
      printed
PASS: AIO-DX -> Standby controller turned off + controllerfs
      add/delete + verify that the execution was interrupted
      and the error message was correctly printed
PASS: AIO-DX -> Standby controller locked-online + controllerfs
      add/delete + verify that the execution was successful
      and the message was correctly printed

Story: 2011117
Task: 50713

Change-Id: Ib12907f1376df67cd7016a0dd90b4d4193a26711
Signed-off-by: Arthur Antonio Duarte Freire <arthurantonio.duartefreire@windriver.com>