Fix wrong mgmt-mac after interface-network-assign
After running system interface-network-assign for the MGMT interface,
the mgmt-mac is not updated in the database and system host-show shows
it as "00:00:00:00:00:00"

There were two checks preventing the database update in case the
bootstrap flag was not set. Both checks are being removed to fix this
issue.

Test Plan:
PASS: SX: system interface-network-assign and check that mgmt_mac is ok
PASS: DX: install system and check that mgmt_mac is ok

Closes-Bug: 2075519

Change-Id: I07a1af3e661d7572e849e75a90979d872e9c9cc5
Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>