Fix 50-remove-calico-ipip.py upgrade script
This script was not taking in account that default-ipv4-ippool does not
exist on a IPv6 setup and therefore it was failing and broke the upgrade
process.

This fix tests the presence of default-ipv4-ippool before using it.

Test Plan:
PASS: execute the script on a IPv4 lab (IPIP disabled)
PASS: execute the script on a IPv6 lab (no errors)

Story: 2011124
Task: 50717

Change-Id: I59ccaad63c388fed67682571ff9f166647d8c27c
Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>