Fix activate script db connection
This commit is to fix the 98-update-isystem-data.py script database
connection issue introduced in [1].

During the activate stage, the database doesn't allow the peer authentication.
To connect to the database, the username/password is required.

[1]  https://review.opendev.org/c/starlingx/config/+/924072

Test Plan:

PASS: build and deploy iso on SX
PASS: run the USM upgrade to next release

Task: 50555
Story: 2010676
Change-Id: I603f1c2adb17c027b6094526e6f010ffdcdefb7d
Signed-off-by: junfeng-li <junfeng.li@windriver.com>