Require OAM to be dual-stack if cluster-pod is dual-stack
It is observed a failure for the pods (when IPv4 is the secondary
address pool) to resolve 169.254.1.1, the gateway used inside to
access the IPv4 network outside. This is typically when OAM subnet
is single-stack IPv6.

When Pod do ARP request for gateway address, host-side calico
interface, which are proxy-arp enabled, respond to ARP request,
if proxy-arp requirement is met: IP to which it has some route.
In case of single-stack IPv6 OAM, host don't have any route
matching to IP 169.254.1.1, leading this issue.

As Pod to internet IPv4 access will require default IPv4 route
in host anyway, which would be provided by OAM, as solution we
are forcing OAM as dual-stack whenever cluster pod subnet is
dual-stack.

Test Plan:
 PASS: AIO-SX dual-stack to single-stack migration:
       - deassigning secondary OAM pool (network-addrpool-remove) before
         deassigning secondary cluster-pod pool, throws error.
       - deassigning secondary OAM pool after deassigning secondary
         cluster-pod pool succeed.

 PASS: AIO-SX single-stack to dual-stack migration:
       - assigning secondary cluster-pod pool (network-addrpool-assign)
         before assigning secondary OAM pool, throws error.
       - assigning secondary cluster-pod pool after secondary OAM pool
         succeed.

Story: 2011027
Task: 50740

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/925604
Change-Id: I958ac51a67234315691753bb0c480c224865e40b
Signed-off-by: Tara Nath Subedi <tara.subedi@windriver.com>