Fix ceph mon and ceph filesystem hiera data
The platform::ceph::params::mon_lv_size parameter was being added to the
host hiera data for all hosts, but should only be added on hosts with
ceph-mon configured.

This also fixes the filesystem ceph that was being added on all
workers for the AIO-DX+ setup, but should only be automatically
configured on the controllers.

Test Plan:
  PASS: Install AIO-DX with Ceph configured
  PASS: Install AIO-DX+ with Ceph configured
  PASS: Install AIO-SX with Ceph configured
  PASS: Install STD with Ceph configured

Closes-bug: 2076369

Depends-on: https://review.opendev.org/c/starlingx/stx-puppet/+/926008

Change-Id: Icb789d8f8778c956bb062280a10f854c971df004
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>