Fix "sysinv-app query <k8s-target-version>"
The "sysinv-app query" command is intended to be used in the upgrade
process, more precisely in the activate process within script 65.
Basically, it collects all installed apps and checks if there is a
tarball with a higher version of the application so that the upgrade
can be performed.

This check is done by consulting the kube_app_bundle table that stores
all the tarballs present in the /usr/local/share/applications/helm/
folder when the system is started.

However, not all applications follow this flow as is the example of
third-party applications. In these cases, we will simply escape this
check.

Along with this change, one more detail was added to script 65. To
prevent the logs of the "sysinv-app query <k8s-target-version>"
command from being logged to stdout, this flag was added:
"--log-file /var/log/software.log".

Test Plan:
PASS: build-pkgs && build-image
PASS: Install wr-analytics on stx-9
PASS: upgrade from stx-9 to master
PASS: 65-k8s-app-upgrade.sh script runs without errors

Closes-bug: 2076911

Change-Id: I2fdff575f767d7cb1e9276cd71d66a30c114929f
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>