Improve adding OSDs to apply puppet manifest
This change was necessary because the check validating if the
config_dict is a duplicate action is not sufficient when the
dict has multiple keys when adding OSDs, so the puppet
manifest does not execute some actions when adding many OSDs
at the same time.
Example dicts:
{
    "personalities": "controller",
    "host_uuids": [
        "a9d2d372-775b-43ea-ad66-ab7807c6f0a6"
    ],
    "stor_uuid": "a095068b-8819-4799-9f7e-cb2e9657e81d",
    "classes": [
        "platform::ceph::runtime_osds"
    ],
    "report_status": "ceph_osd"
}
{
    "personalities": "controller",
    "host_uuids": [
        "a9d2d372-775b-43ea-ad66-ab7807c6f0a6"
    ],
    "stor_uuid": "cb2f8d11-6647-40d3-b112-d9f4b7fca7c1",
    "classes": [
        "platform::ceph::runtime_osds"
    ],
    "report_status": "ceph_osd"
}

To solve this issue, an improvement in checking has been added to
check all keys and values to check whether the action is duplicate or
not.

Test Plan:
 - PASS: Install a Standard system with 4 OSDs added at the same time
 - PASS: Run parallel route config to ensure no stuck 250.001 in alarms

Closes-Bug: 2077120

Change-Id: Ie5a0d4098d6956260cd8ee96ac30a189320e3056
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>