Update status when all ceph mon hosts finish configuration.
This method would run only on controllers that had a valid state and
after running on both, it would automatically give the state of
configured to the worker node. This would cause a problem when the
one of the controllers would have an invalid, changed state when the
method was running, because it would cause the method to indicate
that the config did not run successfully and it would not change the
status of the worker's ceph-mon on the sysinv db, even if it was up and
healthy.

This change makes the method run on all hosts that have a configured
ceph-mon to address each host individually, so the sysinv db indicates
correctly the status of each ceph-mon. This method now checks each
individual host and checks if every task on every host is changed to
configured. If it is configured, the task becomes None, otherwise, if
there are still hosts that are not configured, the state of the
ceph_mon on the host that it is checking will be changed to
configured, but it won't be finished.

Test Plan:
pass: Install Standard config and check if ceph is running
successfully.
PASS: Install Standard config and change the status on
controller-1 while ceph is being configured and check if
ceph-mon-list is correct.
PASS: Install AIO-SX and check if ceph is running successfully and
system ceph-mon-list shows ceph-mon CONFIGURED.
pass: Install AIO-DX and check if ceph is running successfully
and system ceph-mon-list shows ceph-mons CONFIGURED.

Closes-Bug: 2073179

Change-Id: Id672a8362b567f7504beb918105bcb600394873f
Signed-off-by: Luiz Felipe Kina <LuizFelipe.EiskeKina@windriver.com>