Update Postgres Auth and Password Encryption
In PostgreSQL 13, the default authentication and
password encryption method is 'md5'. To enhance
security, both methods have been updated to
'scram-sha-256' in the new software version.

During the USM upgrade process, with the changes
introduced at the 'deploy start' stage, new databases
and roles are created using the authentication and
password encryption protocols required by the new
software version.

This change introduces an upgrade script for
roll-back purposes. In case of a roll-back, the
encryption method used to store the passwords in
the database must be restored.

To ensure successful authentication, both the
password encryption and authentication methods
must be consistent. Therefore, roles must be updated
and saved using the password encryption method
from the previous software version (md5).

Test Plan:
- PASS Fresh Install SX env
   * Verify system status unlock/available

   * Login as admin user in psql
     (psql -U admin -h 127.0.0.1 -d sysinv)
   * Check postgres authorization configuration
     (SELECT * from pg_hba_file_rules;)
   * Check postgres password encryption configuration
     (SELECT rolname, rolpassword
      FROM pg_authid WHERE rolpassword IS NOT NULL;).

- PASS Fresh Install DX env
   * Verify system status unlock/available

   * Login as admin user in psql
     (psql -U admin -h 127.0.0.1 -d sysinv)
   * Check postgres authorization configuration
     (SELECT * from pg_hba_file_rules;)
   * Check postgres password encryption configuration
     (SELECT rolname, rolpassword
      FROM pg_authid WHERE rolpassword IS NOT NULL;).

   * Host swact to controller-1

   * Login as admin user in psql
     (psql -U admin -h 127.0.0.1 -d sysinv)
   * Check postgres authorization configuration
     (SELECT * from pg_hba_file_rules;)
   * Check postgres password encryption configuration
     (SELECT rolname, rolpassword
      FROM pg_authid WHERE rolpassword IS NOT NULL;).

   * collect logs (collect)
   * verify '/var/extra/database/' content

- PASS Upgrade SX
- PASS Upgrade SX-rollback
- PASS Upgrade DX
- PASS Upgrade DX-rollback

Partial-bug: 2069842

Depends-On: https://review.opendev.org/c/starlingx/integ/+/922346

Change-Id: I81bbc880194c2781b32b326e849fce060f69e136
Signed-off-by: Jorge Saffe <jorge.saffe@windriver.com>