Improve certificate deployment by cert-mon
While performing upgrade with https disabled on the system and
during the upgrade activate step,the upgrade script
"41-create-required-platform-certs.py" deploys the
"system-restapi-gui-certificate" certificate,the cert-mon service
enables the "https" first which deploys the "self-signed" ssl
certificate. Simultaneously, if another thread of cert-mon tries
to update the other platform certificates like "docker_registry"
the rest-api call throws time-out error due to the ssl certificate
being updated on the sysinv api, this fix changes the behavior of
"system-restapi-gui-certificate" update by cert-mon by
updating the certificate first and then enabling "https",this
fix also adds lock on method "update_platform_certificate"
which synchronizes the platform certificate deployment and
avoids time-out's.

TestCases:
PASS: Perform an upgrade with https disabled on the system, during
      upgrade activate step, notice that there are no time-out
      errors in the cert-mon log and also observer that there are
      no certificate errors while executing the upgrade activate
      scripts and activation is successful.
PASS: Built ISO with changes, install it and verify that all the
      platform certificates are successfully created.
PASS: Enable https on the system and perform upgrade and verify
      that upgrade activation is successful

Closes-bug: 2077580

Change-Id: I39b8eaee069e02faacfb03a77a66cd57b82a32bb
Signed-off-by: amantri <ayyappa.mantri@windriver.com>