Implement token cache for cert-mon and cert-alarm
These 2 services have some caching in place already but there are
still 3 functions where the token is not cached. Also, the cert-mon
process had 2 versions of the cache object.

This change implements keystone caching for all function calls
in cert-mon and cert-alarm and consolidates caching to a single
object, which will result in greater cache hit and lower miss.

Test case:
PASS: Full build and install
PASS: Enable keystone access log and verify that only 1
      /auth/token request is sent per hour by each service.
PASS: Enable debug logs in for cert-mon and cert-alarm and verify
      that logs show for both services:
      - 'Acquiring new token' once when service starts
      - 'Token is still valid, reusing' for about 1 hour
      - 'Acquiring new token' when it's past 1 hour and
      a new token needs to be requested.

Closes-Bug: 2077662
Change-Id: I6d73a06cf8956604131f8995d9c51ed1d6192f8e
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>