Wait for nginx and cert-manager pods to be ready
Wait for nginx-ingress-controller and cert-manager pods to be ready
during platform upgrades when running script the 65-k8s-app-upgrade.sh
script.

Waiting for those pods is required because both apps need to be fully
running so that the next upgrade steps are successful.

The mechanism for updating fluxCD manifests is done in cycles/routines
that we call reconciliation and applications use 1 minute as the
default. To ensure that the status of pods of version (N+1) is checked
instead of N, a 1-minute delay was added.

Test Plan:
PASS: Upgrade from previous release to mainline
      Confirm that script 65-k8s-app-upgrade.sh finishes only when the
      required pods are ready.

Closes-bug: 2077758

Change-Id: Ie44a2f621f9b7b837be801867f5f30205a7ff68b
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>