Wait reconfiguration in upgrade script #41
Certificate installation triggered by the script
41-create-required-platform-certs.py start several configuration
procedures in sysinv.

To avoid interfering with posterior scripts, this review adds a
verification waiting for out-of-date alarms to clear at the end of the
script.

The proposed wait time is 30 seconds * 40 attempts, which would lead to
a maximum wait of around 20 minutes. After this, we ignore the alarm and
move on as we are not sure of the cause and cert-mon should recover by
itself.

Test plan:
PASS: Perform AIO-SX upgrade. Observe that the script #41 waits for the
      alarm to clear before exiting.
      (three executions, worst took 18 attempts * 30s, around 9 min.)

Story: 2009811
Task: 50914

Change-Id: Ic178608e9bc2e7dcf3a342f008bcd82ff0abc432
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>