Update cert SANs only after initial config
Locks the logic for updating cert SANs when updating OAM to only after
initial config. There could be a case where Kubernetes is not up
during bootstrap when the logic to update the certificate SANs is
called, although this is rare.

Testing
PASS: Bootstrap a multi node system
PASS: Add an IPv6 OAM to an IPv4 multi node system, check system
      commands and software list works over publicURL.

Partial-Bug: 2077671
Change-Id: I74a9ff9e5ae07b52127d7eb7928a7241d510d2a7
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>