Provide app dependency behaviour for deploy activate
The upgrade script 65 has been updated to allow apps to be
installed in parallel to reduce the activate process time.
However, some apps suffer from race conditions due to
dependencies that other apps need to be installed first.

This fix allows apps that have dependencies to be installed
respecting the priority queue, while apps that do not have
dependencies continue to be installed in parallel.

I also take advantage of these changes to improve the function
that checks the status of the ingress-nginx-controller and
cert-manager pods.

A fix was made to the check_pod_readiness function in script
65. The --kubeconfig flag was added to the kubectl command.

Test Plan:
PASS: build-pkgs && build-image
PASS: PASS: upgrade from stx-9 to master
PASS: 65-k8s-app-upgrade.sh script runs without errors
PASS: Applications that can be updated in parallel are updated
      in parallel.
PASS: Applications that need to be serially updated are serially
      updated.

Closes-bug: 2078948

Change-Id: Id7527eff4043ab455e33a531ebb3d486669e1372
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>