Prevent Kube_app Crash Due to NoneType Variable
In one upgrade test, the conditions list was unexpectedly
None, causing the Python code to crash.

Although this issue only occurred once, it is recommended
to implement a check to determine if the object is None
before attempting to iterate over it.

The exact cause of the conditions list being None could
not be determined, as the issue has not reoccurred.

The code will recheck the POD status, so no regressions
are expected.

Test Plan:
PASS: build-pkgs && build-image
PASS: fresh install AIO-SX
PASS: fresh install AIO-DX

Co-Authored-By: Fabiano Correa Mercer
<fabiano.correamercer@windriver.com>

Change-Id: I9d3e312466d21140e433f316c31ad7ba3add68eb
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>