Fix #41 script for upgrade rollback
During rollback, from_release is '24.09' (stx 10), the inverse of what
is during upgrade activate. Fixed script to use the correct value.

Also, systems upgraded from stx 9 to stx 10 with HTTPS disabled have
an SSL cert auto generated during the upgrade (issued by
system-local-ca), in order to enable HTTPS.

When performing rollback, this certificate is not present in the db
and HTTPS is disabled again in stx 9, but the actual file is still
present under '/etc/ssl/private'. The file is not used anymore, but
it still shows when the script 'show-certs.sh' is executed.

This review removes it as to achieve the previous stx 9 configuration
after the upgrade rollback. HTTPS is disabled beforehand to avoid
issues.

Test plan:
PASS: With AIO-SX with HTTPS disabled, perform stx 9 to stx 10
      upgrade until activation is done (HTTPS cert is auto generated).
      Abort the upgrade and perform rollback procedure. Observe that
      the auto generated HTTPS cert is deleted and is not present in
      stx 9 side.

Story: 2009811
Task: 50963

Change-Id: I84b08811a60cb4deccfa159895658b67427b3868
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>