Health check in k8s commands for #41 upgrade script
The script 41-create-required-platform-certs.py is currently the first
to attempt requests to kubeapi server, which is being unresponsive in
some attempts to perform upgrade activation after a swact.

Also, cert-mon will install a ssl_ca cert in the background when the
'ca.crt' field of 'system-local-ca' secret changes. The resulting
reconfiguration can be also a cause of the issues when calling kubeapi
server.

This modification adds health checks before each kubeapi server call in
this script, and also a conditional wait for the system reconfiguration
after 'system-local-ca' is changed.

Test plan:
PASS: Perform the upgrade activation.
PASS: Perform the upgrade activation, but manually kill the kubeapi
      server process before the #41 script is called. Observed that the
      script detects the issue and retries. Kubeapi server recovered
      and the script proceeded.

Story: 2010676
Task: 50970

Change-Id: I6e0da85401aee3609d5c2c62a5f9d6a7a97670b4
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>