Suppress misleading errors messages on script 65
This commit fixes a minor issues with script 65, where uses
grep to check if an app is already present in the upgrading
apps control file, but this file does not exist. In this case,
grep produces an error message that does not make the script to
fail, but may lead the user to believe it did, so, this message
is suppressed by this commit, by letting grep execute only if
the file exists.

Test Plan
PASS: AIO-SX - run activate successfully using USM stx-8 -> stx-10
PASS: run the script manually while the file does not exist
PASS: run the script manually while the file exists

Story: 2010676
Task: 50975

Change-Id: Iaf4b55cb3947cfa887e0a5677dca4f61f8fe078c
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>